requirements.txt
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
pymongo>=4.0.0
openpyxl>=3.0.0
mysql-connector-python>=8.0.0
